rename rule to async-contains-then-or-catch
