+++ docs/rules/async-contains-then-or-catch.md
