+++ lib/rules/async-contains-then-or-catch.js
